add importing product by EAN

+++ Backend/StockWise/Dtos/CreateProductDto.cs
@@ -6,8 +6,8 @@
         public string EAN { get; set; }
         public string? Image { get; set; }
         public string Description { get; set; }
-        public double ShoppingPrice { get; set; }
-        public double SellingPrice { get; set; }
+        public decimal ShoppingPrice { get; set; }
+        public decimal SellingPrice { get; set; }
         public string Category { get; set; }
     }
 }
